+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.1] - 2024-08-24
+### Added
+- N/A
+
+### Changed
+- N/A
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- Fix for `Collection::remove` method to properly remove `Aurora` models
+- Fix for `Aurora::update` method to ensure deleted relations are properly deleted from the database
+
+### Security
+- N/A
+
+---
+
 ## [2.7.0] - 2024-08-23
 ### Added
 - Implemented new `isEmpty` method on `Collection`